LOW-1149: SKB support
Add support for checking of SKBs. For now there is just a single test, which tests that SKBs are not going stale.

Because adding SKB support required adding a whole new Jira query, I split this out into a separate tests/skbs folder, and moved the project-specific tests into a tests/project folder.
There's some duplication of code across the two folders; that can wait until there is clarity on how best to structure this.